chore: refactor useSpaceSpeaker hooks into different ones
