+++ pho/.eslintrc
@@ -21,6 +21,7 @@
   },
   "plugins": ["react", "prettier"],
   "rules": {
+    "@typescript-eslint/no-dynamic-delete": "off",
     "@typescript-eslint/strict-boolean-expressions": "off",
     "@typescript-eslint/explicit-function-return-type": "off",
     "no-extra-semi": "off",
